Edit app/src/main/java/com/example/widgets/WidgetArt.kt via Bridge

+++ app/src/main/java/com/example/widgets/WidgetArt.kt
@@ -163,7 +163,7 @@ object WidgetArt {
         val tick = Paint().apply { isAntiAlias = true; color = Color.argb(120, 255, 224, 130); strokeWidth = dp(ctx, 2f) }
         for (i in 0 until 12) {
             val a = i * 30f * PI.toFloat() / 180
-            c.drawLine(cx + cos(a).toFloat() * r * 0.78f, cy + sin(a).toFloat() * r * 0.78f, cx + cos(a).toFloat() * r * 0.9f, cy + sin(a).toFloat() * r * 0.9f, tick)
+            c.drawLine(cx + cos(a.toDouble()).toFloat() * r * 0.78f, cy + sin(a.toDouble()).toFloat() * r * 0.78f, cx + cos(a.toDouble()).toFloat() * r * 0.9f, cy + sin(a.toDouble()).toFloat() * r * 0.9f, tick)
         }
         for (i in 0 until 14) {
             val a = (i * 25.7f + frame * 9f) * PI.toFloat() / 180
